analytics/requirements.txt
# requirements.txt

# Data processing and analysis
pandas==1.3.5
numpy==1.21.5
scipy==1.7.3

# Machine Learning
scikit-learn==1.0.2
tensorflow==2.8.0
keras==2.8.0

# Data visualization
matplotlib==3.5.1
seaborn==0.11.2
plotly==5.6.0

# Jupyter and IPython
jupyter==1.0.0
ipython==8.0.1

# Database connections
sqlalchemy==1.4.31
psycopg2-binary==2.9.3

# API and web framework
flask==2.0.2
dash==2.0.0

# Authentication and security
pyjwt==2.3.0
bcrypt==3.2.0

# Blockchain interaction
web3==5.28.0

# Testing
pytest==7.0.1

# Logging and monitoring
logging==0.4.9.6

# Environment management
python-dotenv==0.19.2

# Production server
gunicorn==20.1.0
